fix: course scraper adjustment for "courseblock"  change (#194)
<!--- Provide a general summary of your changes in the Title above -->

## Description

rewrote `scrapeCoursesInDepartment` to parse per-course `.courseblock`
elements and their `.detail-*` children instead of relying on
newline-splitting. This matches the current catalogue HTML and removes
brittleness around formatting changes.

also addressed missing `departmentAlias` in course by adding a mapping
on completed scrape

## Related Issue

addresses #192, #193, #195, #152

## Motivation and Context

course scraper needs to be fixed to run for the 25/26 academic year

## How Has This Been Tested?

current testing status shows that entries for courses are accurately
updated with the exception of `shortened_dept` and `department_alias`.
shortened dept is addressed when running migration 10, but the setting
of `department_alias` needs further investiagion

## Screenshots (if appropriate):

## Types of changes

- [X] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [ ] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.

---------

Co-authored-by: Sanskar Mishra <sanskarmishra@Sanskars-MacBook-Pro-781.local>
Co-authored-by: Sanskar Mishra <sanskarmishra@mac.lan>