fix: unified search results of different types can't clobber each other (#189)
## Description

By asking the database to return the type of a search result, we can
discern which table the result came from (courses or instructors).
We then use this to look the ID up in the correct table.
Since we are now using two distinct maps keyed by ID (one for course
IDs, one for instructor IDs), it is now possible to return two entities
of different types with identical ID.

This also has the side effect of reducing failed ID lookups since we
never look up an instructor ID as a course or vice versa.

## Related Issue

By the third statement above, we fix issue #188.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [ ] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.

---------

Co-authored-by: Sanskar Mishra <69398106+sanskarm7@users.noreply.github.com>