feat(degreeworks-scraper): 2025 updates (#223)
Closes #211.

This meta-PR closes the sub-issues described on that issue.
Each meaningful commit on this PR is a squash merge of one of the
sub-PRs.
Progress is best followed on those PRs.
To summarize here:

* #213: We optimize major scraping using a table found on the Catalogue.
* #214: We correctly infer the most recent catalogue year.
* #218: We account for the loss of information from DegreeWorks about
associations between specializations and majors by inferring this
information ourselves. This also lets us discover previously uncollected
data on specializations for majors where specializations are optional.
* #222: We learn to fetch requirements from schools (except Humanities)
and provide these along with the majors to which they pertain.

We also add Zod v4 in the DegreeWorks scraper as a limited pilot in
pursuit of eventually migrating the entire project.

## Types of changes

- [x] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [x] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.