feat(programs): 4-year CHC requirements (#183)
## Description

Allow the collection of the 4-year CHC program block, if found.
As mentioned on the issue, we must be scraping from an account in the
four-year CHC program.
As of yet we do not actually know whether the requirements block we
obtain is for four-year CHC versus two-year CHC or another variant.
We need to see both blocks side-by-side to learn how to tell them apart.

~~**Blocked by #181.**~~

## Related Issue

#182 

## Types of changes

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:

- [x] My code involves a change to the database schema.
- [ ] My code requires a change to the documentation.